refactor: class 명 변경 및 계층 이동
