+++ backend/src/main/java/com/ice/studyroom/domain/membership/application/MembershipService.java
@@ -15,11 +15,10 @@ import com.ice.studyroom.domain.identity.domain.service.TokenService;
 import com.ice.studyroom.domain.identity.infrastructure.security.JwtTokenProvider;
 import com.ice.studyroom.domain.membership.domain.entity.Member;
 import com.ice.studyroom.domain.membership.domain.service.MemberDomainService;
-import com.ice.studyroom.domain.membership.domain.service.MemberEmailService;
 import com.ice.studyroom.domain.membership.domain.vo.Email;
 import com.ice.studyroom.domain.membership.infrastructure.persistence.MemberRepository;
 import com.ice.studyroom.domain.membership.presentation.dto.request.MemberCreateRequest;
-import com.ice.studyroom.domain.membership.presentation.dto.request.MemberEmailRequest;
+import com.ice.studyroom.domain.membership.presentation.dto.request.EmailVerificationRequest;
 import com.ice.studyroom.domain.membership.presentation.dto.request.MemberEmailVerificationRequest;
 import com.ice.studyroom.domain.membership.presentation.dto.request.MemberLoginRequest;
 import com.ice.studyroom.domain.membership.presentation.dto.request.TokenRequest;
@@ -39,7 +38,7 @@ public class MembershipService {
 	private final JwtTokenProvider jwtTokenProvider;
 	private final TokenService tokenService;
 	private final AuthenticationManager authenticationManager;
-	private final MemberEmailService memberEmailService;
+	private final EmailVerificationService memberEmailService;
 
 	public MemberResponse createMember(MemberCreateRequest request) {
 		memberDomainService.validateEmailUniqueness(Email.of(request.email()));
@@ -86,7 +85,7 @@ public class MembershipService {
 		return MemberResponse.of("success");
 	}
 
-	public MemberEmailResponse sendMail(MemberEmailRequest request) {
+	public MemberEmailResponse sendMail(EmailVerificationRequest request) {
 		memberDomainService.validateEmailUniqueness(Email.of(request.email()));
 		memberEmailService.sendCodeToEmail(request.email());
 		return MemberEmailResponse.of("인증 메일이 전송되었습니다.");
