+++ backend/src/main/java/com/ice/studyroom/domain/membership/presentation/dto/request/EmailVerificationRequest.java
@@ -3,7 +3,7 @@ package com.ice.studyroom.domain.membership.presentation.dto.request;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public record MemberEmailRequest(
+public record EmailVerificationRequest(
 	@NotBlank(message = "이메일은 필수입니다.")
 	@Email(message = "올바른 이메일 형식이 아닙니다.")
 	String email
